Modified   include/Common.h Modified   include/Controls/Potentiometer.h Modified   include/DSP/AutoWah.h Modified   include/DSP/Chorus.h Modified   include/DSP/Reverb.h Modified   include/DSP/Test.h Modified   include/DSP/Tremolo.h Modified   include/DSP/Wah.h Modified   include/Filters/BandPassFilter.h Modified   include/Filters/BiquadFilter.h Modified   include/Filters/HighPassFilter.h Modified   include/Filters/LowPassFilter.h Modified   include/Filters/NoiseGateFilter.h Missing    include/Filters/ToneControlFilter.h Modified   include/Math.h Added      include/Filters/BandStopFilter.h Added      include/Filters/SingleToneControlFilter.h Added      include/Filters/ThreeToneControlFilter.h

+++ include/DSP/Chorus.h
@@ -99,7 +99,7 @@ private:
 		uint32 indexDelayed = (m_DelayBufferIndex + index) % m_DelayBufferLength;
 		uint32 indexNext = (indexDelayed + 1) % m_DelayBufferLength;
 
-		return Math::Lerp(m_DelayBuffer[indexDelayed], m_DelayBuffer[indexNext], fraction) / MAX_BUFFER_ELEMENT_COEFF;
+		return Math::Lerp((double)m_DelayBuffer[indexDelayed], m_DelayBuffer[indexNext], fraction) / MAX_BUFFER_ELEMENT_COEFF;
 	}
 
 private:
